@@ -26,30 +26,39 @@
 E-ARK : Information package validation
         Command line validation application
 """
-import argparse
-from pprint import pprint
+import json
 import os.path
+from pathlib import Path
 import sys
+from typing import Optional, Tuple
+import importlib.metadata
 
-import eark_validator.structure as STRUCT
+import argparse
 
-__version__ = '0.1.0'
+from eark_validator.model import ValidationReport
+import eark_validator.packages as PACKAGES
+from eark_validator.infopacks.package_handler import PackageHandler
+from eark_validator.specifications.specification import SpecificationVersion
+
+__version__ = importlib.metadata.version('eark_validator')
 
 defaults = {
-    'description': """E-ARK Information Package validation (ip-check).
-ip-check is a command-line tool to analyse and validate the structure and
+    'description': """E-ARK Information Package validation (eark-validator).
+eark-validator is a command-line tool to analyse and validate the structure and
 metadata against the E-ARK Information Package specifications.
 It is designed for simple integration into automated work-flows.""",
     'epilog': """
 DILCIS Board (http://dilcis.eu)
 See LICENSE for license information.
-GitHub: https://github.com/E-ARK-Software/py-rest-ip-validator
-Author: Carl Wilson (OPF), 2020-2023
-Maintainer: Carl Wilson (OPF), 2020-2023"""
+GitHub: https://github.com/E-ARK-Software/eark-validator
+Author: Carl Wilson (OPF), 2020-2024
+Maintainer: Carl Wilson (OPF), 2020-2024"""
 }
 
 # Create PARSER
-PARSER = argparse.ArgumentParser(description=defaults['description'], epilog=defaults['epilog'])
+PARSER = argparse.ArgumentParser(prog='eark-validator',
+                                 description=defaults['description'],
+                                 epilog=defaults['epilog'])
 
 def parse_command_line():
     """Parse command line arguments."""
@@ -57,18 +66,35 @@ def parse_command_line():
     PARSER.add_argument('-r', '--recurse',
                         action='store_true',
                         dest='inputRecursiveFlag',
-                        default=True,
+                        default=False,
                         help='When analysing an information package recurse into representations.')
     PARSER.add_argument('-c', '--checksum',
                         action='store_true',
                         dest='inputChecksumFlag',
                         default=False,
-                        help='Calculate and verify file checksums in packages.')
+                        help='Calculate and verify package checksums.')
+    PARSER.add_argument('-m', '--manifest',
+                        action='store_true',
+                        dest='inputManifestFlag',
+                        default=False,
+                        help='Display package manifest information.')
     PARSER.add_argument('-v', '--verbose',
                         action='store_true',
                         dest='outputVerboseFlag',
                         default=False,
-                        help='report results in verbose format')
+                        help='Verbose reporting for selected output options.')
+    PARSER.add_argument('--schema',
+                        action='store_true',
+                        dest='output_schema',
+                        default=False,
+                        help='Request display of the JSON schema of the output report.')
+    PARSER.add_argument('-s', '--specification_version',
+                        nargs='?',
+                        dest='specification_version',
+                        default=SpecificationVersion.V2_1_0,
+                        type=SpecificationVersion,
+                        choices=list(SpecificationVersion),
+                        help='Specification version used for validation. Default is %(default)s.')
     PARSER.add_argument('--version',
                         action='version',
                         version=__version__)
@@ -89,37 +115,48 @@ def main():
     # Get input from command line
     args = parse_command_line()
     # If no target files or folders specified then print usage and exit
-    if not args.files:
+    if _is_show_help(args):
         PARSER.print_help()
+
+    if args.output_schema:
+        print(json.dumps(ValidationReport.model_json_schema(), indent=2))
+        sys.exit(0)
 
     # Iterate the file arguments
     for file_arg in args.files:
-        _loop_exit, _ = _validate_ip(file_arg)
+        _loop_exit, _ = _validate_ip(file_arg, args.specification_version)
         _exit = _loop_exit if (_loop_exit > 0) else _exit
     sys.exit(_exit)
 
-def _validate_ip(info_pack):
-    ret_stat = _check_path(info_pack)
-    struct_details = STRUCT.validate_package_structure(info_pack)
-    pprint('Path {}, struct result is: {}'.format(info_pack,
-                                                         struct_details.status))
-    for error in struct_details.errors:
-        pprint(error.to_json())
+def _validate_ip(path: str, version: SpecificationVersion) -> Tuple[int, Optional[ValidationReport]]:
+    ret_stat, checked_path = _check_path(path)
+    if ret_stat > 0:
+        return ret_stat, None
+    report = PACKAGES.PackageValidator(checked_path, version).validation_report
+    print(f'Path {checked_path}, struct result is: {report.structure.status.value}')
+    # for message in report.structure.messages:
+    print(report.model_dump_json())
 
-    return ret_stat, struct_details
+    return ret_stat, report
 
-def _check_path(path):
+def _check_path(path: str) -> Tuple[int, Optional[Path]]:
     if not os.path.exists(path):
         # Skip files that don't exist
-        pprint('Path {} does not exist'.format(path))
-        return 1
+        print(_format_check_path_message(path, 'does not exist'))
+        return 1, None
     if os.path.isfile(path):
         # Check if file is a archive format
-        if not STRUCT.ArchivePackageHandler.is_archive(path):
+        if not PackageHandler.is_archive(path):
             # If not we can't process so report and iterate
-            pprint('Path {} is not a file we can process.'.format(path))
-            return 2
-    return 0
+            print(_format_check_path_message(path, 'is not an archive file or directory'))
+            return 2, None
+    return 0, Path(path)
+
+def _format_check_path_message(path: Path, message: str) -> str:
+    return f'Processing terminated, path: {path} {message}.'
+
+def _is_show_help(args) -> bool:
+    return not args.files and not args.output_schema
 
 # def _test_case_schema_checks():
 if __name__ == '__main__':
