@@ -24,17 +24,62 @@
 #
 """Module to capture everything schematron validation related."""
 import os
-from importlib_resources import files
+from urllib.request import urlopen
 from typing import Generator
+
+from importlib_resources import files
 
 from lxml import etree as ET
 from lxml.isoschematron import Schematron
 
-from .resources import schematron as SCHEMATRON
 from eark_validator.const import NO_PATH, NOT_FILE
+from .resources import schematron as SCHEMATRON
+from .resources import vocabs as vocabularies
 
 SCHEMATRON_NS = '{http://purl.oclc.org/dsdl/schematron}'
 SVRL_NS = '{http://purl.oclc.org/dsdl/svrl}'
+
+class SchematronTests():
+    __vocabulary_definitions = {
+        '@TYPE': 'https://earkcsip.dilcis.eu/schema/CSIPVocabularyContentCategory.xml',
+        '@csip:CONTENTINFORMATIONTYPE': 'https://earkcsip.dilcis.eu/schema/CSIPVocabularyContentInformationType.xml',
+        '@csip:OAISPACKAGETYPE': 'https://earkcsip.dilcis.eu/schema/CSIPVocabularyOAISPackageType.xml',
+        '@STATUS': 'https://earkcsip.dilcis.eu/schema/CSIPVocabularyStatus.xml'
+    }
+
+    tests = {}
+
+    def __init__(self):
+        for attribute, vocabulary_uri in self.__vocabulary_definitions.items():
+            self.tests[attribute + '_vocabulary_test'] = self.__create_vocabulary_test(attribute, vocabulary_uri)
+
+        self.tests['@MIMETYPE_IANA_test'] = self.___create_IANA_test()
+
+    def __create_vocabulary_test(self, attribute: str, vocabulary_uri: str) -> str:
+        vocabulary_tests = []
+        for line_bytes in urlopen(vocabulary_uri):
+            line = line_bytes.decode('utf-8')
+            if 'Term' not in line:
+                continue
+
+            start = line.find('>') + 1
+            end = line.find('<', start)
+
+            vocabulary_item = line[start:end]
+            vocabulary_tests.append(f"({attribute} = '{vocabulary_item}')")
+
+        return ' or '.join(vocabulary_tests)
+
+    def ___create_IANA_test(self) -> str:
+        mime_tests = []
+        with open(str(files(vocabularies).joinpath('IANA.txt')), 'r') as iana:
+            for mime_type in iana:
+                mime_type = mime_type.rstrip('\n')
+                mime_tests.append(f"(@MIMETYPE = '{mime_type}')")
+
+        return ' or '.join(mime_tests)
+
+schematron_tests = SchematronTests()
 
 class SchematronRuleset():
     """Encapsulates a set of Schematron rules loaded from a file."""
@@ -45,11 +90,21 @@ class SchematronRuleset():
             raise ValueError(NOT_FILE.format(sch_path))
         self._path = sch_path
         try:
-            self._schematron = Schematron(file=self._path, store_schematron=True, store_report=True)
-        except ET.SchematronParseError as ex:
-            raise ValueError('Rules file is not valid XML: {}. {}'.format(sch_path, ex.error_log.last_error.message ))
+            with open(sch_path) as schematron_file:
+                schematron_data = schematron_file.read()
+                for test_name, test_value in schematron_tests.tests.items():
+                    schematron_data = schematron_data.replace(test_name, test_value)
+
+                tree = ET.XML(schematron_data)
+                self._schematron = Schematron(etree=tree, store_schematron=True, store_report=True)
+        except (ET.SchematronParseError, ET.XMLSyntaxError) as ex:
+            ex_mess = ex.error_log.last_error.message # pylint: disable=E1101
+            subject = 'Schematron'
+            raise ValueError(f'Rules file is not valid {subject}: {sch_path}. {ex_mess}') from ex
         except KeyError as ex:
-            raise ValueError('Rules file is not valid Schematron: {}. {}'.format(sch_path, ex.__doc__))
+            ex_mess = ex.__doc__
+            subject = 'XML'
+            raise ValueError(f'Rules file is not valid {subject}: {sch_path}. {ex_mess}') from ex
 
     @property
     def path(self) -> str:
@@ -61,11 +116,20 @@ class SchematronRuleset():
         """Return the Schematron object."""
         return self._schematron
 
-    def get_assertions(self) -> Generator[ ET.Element, None, None]:
-        """Generator that returns the rules one at a time."""
+    @property
+    def assertions(self) -> Generator[ ET.Element, None, None]:
+        """Generator that returns the assertion rules one at a time."""
         xml_rules = ET.XML(bytes(self.schematron.schematron))
         for ele in xml_rules.iter():
             if ele.tag == SCHEMATRON_NS + 'assert':
+                yield ele
+
+    @property
+    def reports(self) -> Generator[ ET.Element, None, None]:
+        """Generator that returns the report rules one at a time."""
+        xml_rules = ET.XML(bytes(self.schematron.schematron))
+        for ele in xml_rules.iter():
+            if ele.tag == SCHEMATRON_NS + 'report':
                 yield ele
 
     def validate(self, to_validate: str) -> ET.Element:
@@ -74,5 +138,5 @@ class SchematronRuleset():
         self.schematron.validate(xml_file)
         return self.schematron.validation_report
 
-def get_schematron_path(id: str, section: str) -> str:
-    return str(files(SCHEMATRON).joinpath(id).joinpath('mets_{}_rules.xml'.format(section)))
+def get_schematron_path(version: str, spec_id: str, section: str) -> str:
+    return str(files(SCHEMATRON).joinpath(version).joinpath(spec_id).joinpath(f'mets_{section}_rules.xml'))
