@@ -25,264 +25,305 @@
 """Module covering tests for package structure errors."""
 import os
 import unittest
+from pathlib import Path
 
 from eark_validator import structure as STRUCT
-from eark_validator.rules import Severity
+from eark_validator.model import Severity
 from tests.utils_test import contains_rule_id
 
-EXP_NOT_WELLFORMED = 'Expecting status NotWellFormed, not {}'
+EXP_NOT_WELLFORMED = 'Expecting status NOTWELLFORMED, not {}'
 EXP_WELLFORMED = 'Expecting status WellFormed, not {}'
 EXP_ERRORS = 'Expecting {} errors but found {}'
 class StructValidationTests(unittest.TestCase):
     """Unit tests covering structural validation of information packages, spcifically
     unpacking archived packages and establishing that the files and folders specified
     if the CSSTRUCT are present."""
+    ip_res_root: str = os.path.join(os.path.dirname(__file__), 'resources', 'ips')
+    bad_ip_root: str = os.path.join(ip_res_root, 'bad')
 
-    def test_check_package_root_single(self):
-        """Dedicated test for package root detection errors."""
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'unpacked',
-                               'single_file')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.NotWellFormed,
-                        EXP_NOT_WELLFORMED.format(details.status))
-        val_errors = details.errors
-        err_count = 1
-        self.assertTrue(len(val_errors) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR4',
-                                         severity=Severity.ERROR))
-
-    def test_check_package_root_multi_dir(self):
-        """Dedicated test for package root detection errors."""
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'unpacked',
-                               'multi_dir')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.NotWellFormed,
-                        EXP_NOT_WELLFORMED.format(details.status))
-        val_errors = details.errors
-        err_count = 1
-        self.assertTrue(len(val_errors) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR4',
-                                         severity=Severity.ERROR))
-
-    def test_check_package_root_multi_file(self):
-        """Dedicated test for package root detection errors."""
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'unpacked',
-                               'multi_file')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.NotWellFormed,
-                        EXP_NOT_WELLFORMED.format(details.status))
-        val_errors = details.errors
-        err_count = 1
-        self.assertTrue(len(val_errors) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR4',
-                                         severity=Severity.ERROR))
-
-    def test_check_package_root_multi_var(self):
-        """Dedicated test for package root detection errors."""
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'unpacked',
-                               'multi_var')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.NotWellFormed,
-                        EXP_NOT_WELLFORMED.format(details.status))
-        val_errors = details.errors
-        err_count = 1
-        self.assertTrue(len(val_errors) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR4',
-                                         severity=Severity.ERROR))
-
-    def test_single_file_archive(self):
+    def test_str1_bad_path(self):
         """Test a package that's just a compressed single file."""
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'struct',
-                               'empty.zip')
-        details = STRUCT.validate_package_structure(ip_path)
-        val_errors = details.errors
+        ip_path = Path(os.path.join(os.path.dirname(__file__), 'resources', 'empty.file'))
+        _, details = STRUCT.validate(ip_path)
         err_count = 1
-        self.assertTrue(len(val_errors) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR4',
+        self.assertEqual(len(details.errors), err_count,
+                        EXP_ERRORS.format(err_count, len(details.errors)))
+        self.assertTrue(contains_rule_id(details.errors, 'CSIPSTR1',
                                          severity=Severity.ERROR))
+
+    def test_str1_package_root_single(self):
+        """Dedicated test for package root detection errors."""
+        ip_path = Path(os.path.join(self.bad_ip_root, 'single_file.zip'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.NOTWELLFORMED,
+                        EXP_NOT_WELLFORMED.format(details.status))
+        err_count = 1
+        self.assertEqual(len(details.errors), err_count,
+                        EXP_ERRORS.format(err_count, len(details.errors)))
+        self.assertTrue(contains_rule_id(details.errors, 'CSIPSTR1',
+                                         severity=Severity.ERROR))
+
+    def test_str1_package_root_multi_dir(self):
+        """Dedicated test for package root detection errors."""
+        ip_path = Path(os.path.join(self.bad_ip_root, 'multi_dir'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.NOTWELLFORMED,
+                        EXP_NOT_WELLFORMED.format(details.status))
+        err_count = 1
+        self.assertEqual(len(details.errors), err_count,
+                        EXP_ERRORS.format(err_count, len(details.errors)))
+        self.assertTrue(contains_rule_id(details.errors, 'CSIPSTR1',
+                                         severity=Severity.ERROR))
+
+    def test_str1_package_root_multi_file(self):
+        """Dedicated test for package root detection errors."""
+        ip_path = Path(os.path.join(self.bad_ip_root, 'multi_file.zip'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.NOTWELLFORMED,
+                        EXP_NOT_WELLFORMED.format(details.status))
+        err_count = 1
+        self.assertEqual(len(details.errors), err_count,
+                        EXP_ERRORS.format(err_count, len(details.errors)))
+        self.assertTrue(contains_rule_id(details.errors, 'CSIPSTR1',
+                                         severity=Severity.ERROR))
+
+    def test_str1_package_root_multi_var(self):
+        """Dedicated test for package root detection errors."""
+        ip_path = Path(os.path.join(self.bad_ip_root, 'multi_var.zip'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.NOTWELLFORMED,
+                        EXP_NOT_WELLFORMED.format(details.status))
+        err_count = 1
+        self.assertEqual(len(details.errors), err_count,
+                        EXP_ERRORS.format(err_count, len(details.errors)))
+        self.assertTrue(contains_rule_id(details.errors, 'CSIPSTR1',
+                                         severity=Severity.ERROR))
+
+    def test_str1_single_file_archive(self):
+        """Test a package that's just a compressed single file."""
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'empty.zip'))
+        _, details = STRUCT.validate(ip_path)
+        err_count = 1
+        self.assertEqual(len(details.errors), err_count,
+                        EXP_ERRORS.format(err_count, len(details.errors)))
+        self.assertTrue(contains_rule_id(details.errors, 'CSIPSTR1',
+                                         severity=Severity.ERROR))
+
+    def test_no_messages(self):
+        """Test package with no METS.xml file"""
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_messages.tar.gz'))
+        is_valid, details = STRUCT.validate(ip_path)
+        self.assertTrue(is_valid)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_NOT_WELLFORMED.format(details.status))
+        self.assertEqual(len(details.messages), 0)
 
     def test_minimal(self):
         """Test minimal STRUCT with schemas, the basic no errors but with warnings package."""
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'minimal',
-                               'minimal_IP_with_schemas.zip')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.WellFormed,
+        ip_path = Path(os.path.join(self.ip_res_root, 'minimal',
+                               'minimal_IP_with_schemas.zip'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
                         EXP_WELLFORMED.format(details.status))
-        val_errors = details.warnings
-        self.assertTrue(len(val_errors) == 3,
-                        'Expecting 3 errors but found {}'.format(len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR12',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR13',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR15',
-                                         severity=Severity.WARN))
+        val_warnings = details.warnings
+        self.assertEqual(len(val_warnings), 5,
+                        'Expecting 2 warnings but found {}'.format(len(val_warnings)))
+        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR6',
+                                         severity=Severity.WARNING))
+        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR7',
+                                         severity=Severity.WARNING))
+        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR12',
+                                         severity=Severity.WARNING))
+        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR13',
+                                         severity=Severity.WARNING))
+        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR16',
+                                         severity=Severity.WARNING))
 
-    def test_nomets(self):
+    def test_str3_package(self):
+        """Test minimal STRUCT with schemas, the basic no errors but with warnings package."""
+        ip_path = Path(os.path.join(self.ip_res_root, 'unpacked',
+                               '733dc055-34be-4260-85c7-5549a7083031'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_WELLFORMED.format(details.status))
+        self.assertEqual(len(details.warnings), 1,
+                        'Expecting 1 warning but found {}'.format(len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR16',
+                                         severity=Severity.WARNING))
+        self.assertEqual(len(details.infos), 2,
+                        'Expecting 2 info messages but found {}'.format(len(details.infos)))
+        self.assertTrue(contains_rule_id(details.infos, 'CSIPSTR3',
+                                         severity=Severity.INFORMATION))
+
+    def test_str4_nomets(self):
         """Test package with no METS.xml file"""
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'struct',
-                               'no_mets.tar.gz')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.NotWellFormed,
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_mets.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.NOTWELLFORMED,
                         EXP_NOT_WELLFORMED.format(details.status))
-        val_errors = details.errors
         err_count = 1
-        self.assertTrue(len(val_errors) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR4'))
+        self.assertEqual(len(details.messages), err_count,
+                        EXP_ERRORS.format(err_count, len(details.messages)))
+        self.assertTrue(contains_rule_id(details.messages, 'CSIPSTR4'))
 
-        val_warnings = details.warnings
-        self.assertTrue(len(val_warnings) == 3,
-                        'Expecting 3 errors but found {}'.format(len(val_warnings)))
-        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR12',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR13',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR15',
-                                         severity=Severity.WARN))
-
-    def test_nomd(self):
+    def test_str5_nomd(self):
         # test as root
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'struct',
-                               'no_md.tar.gz')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.WellFormed,
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_md.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
                         EXP_WELLFORMED.format(details.status))
-        val_errors = details.warnings
-        err_count = 4
-        self.assertTrue(len(val_errors) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR5',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR12',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR13',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR15',
-                                         severity=Severity.WARN))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR5',
+                                         severity=Severity.WARNING))
 
-    def test_noschema(self):
+    def test_str6_nopres(self):
         # test as root
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'struct',
-                               'no_schemas.tar.gz')
-        details = STRUCT.validate_package_structure(ip_path)
-
-        self.assertTrue(details.status == STRUCT.StructureStatus.WellFormed,
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_pres.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
                         EXP_WELLFORMED.format(details.status))
-        val_warnings = details.warnings
-        for entry in val_warnings:
-            print(str(entry))
-        err_count = 4
-        self.assertTrue(len(val_warnings) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_warnings)))
-        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR12',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR13',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR15',
-                                         severity=Severity.WARN))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR6',
+                                         severity=Severity.WARNING))
 
-    def test_nodata(self):
+    def test_str7_nodesc(self):
         # test as root
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'struct',
-                               'no_data.tar.gz')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.WellFormed,
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_desc.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
                         EXP_WELLFORMED.format(details.status))
-        val_errors = details.warnings
-        err_count = 4
-        self.assertTrue(len(val_errors) == err_count,
-                        EXP_ERRORS.format(err_count, len(val_errors)))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR11',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR12',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR13',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR15',
-                                         severity=Severity.WARN))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR7',
+                                         severity=Severity.WARNING))
 
-    def test_noreps(self):
-        ip_path = os.path.join(os.path.dirname(__file__), 'resources', 'ips', 'struct',
-                               'no_reps.tar.gz')
-        details = STRUCT.validate_package_structure(ip_path)
-        self.assertTrue(details.status == STRUCT.StructureStatus.WellFormed,
+    def test_str8_noother(self):
+        # test as root
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_other.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_WELLFORMED.format(details.status))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.infos)))
+        self.assertTrue(contains_rule_id(details.infos, 'CSIPSTR8',
+                                         severity=Severity.INFORMATION))
+
+    def test_str9_noreps(self):
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_reps.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
                         EXP_WELLFORMED.format(details.status))
         val_warnings = details.warnings
-        print('ERRORS')
-        for err in details.messages:
-            print(err)
         err_count = 1
-        self.assertTrue(len(val_warnings) == err_count,
+        self.assertEqual(len(val_warnings), err_count,
                         EXP_ERRORS.format(err_count, len(val_warnings)))
         self.assertTrue(contains_rule_id(val_warnings, 'CSIPSTR9',
-                                         severity=Severity.WARN))
-    """Unit tests covering structural validation of information packages, spcifically
-    unpacking archived packages and establishing that the files and folders specified
-    if the CSIP are present."""
+                                         severity=Severity.WARNING))
 
-    def test_manifest_nomets(self):
-        """Ensure proper behaviour when no METS file is present."""
+    def test_str10_emptyreps(self):
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'empty_reps.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_WELLFORMED.format(details.status))
+        err_count = 1
+        self.assertEqual(len(details.messages), err_count,
+                        EXP_ERRORS.format(err_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR10',
+                                         severity=Severity.WARNING))
+
+    def test_str11_nodata(self):
         # test as root
-        man_no_mets = STRUCT.StructureChecker('no_mets', has_mets=False)
-        val_errors = man_no_mets.validate_manifest()
-        self.assertTrue(len(val_errors) == 1)
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR4'))
-        val_errors = man_no_mets.validate_manifest(is_root=False)
-        self.assertTrue(len(val_errors) == 2)
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR12',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR11',
-                                         severity=Severity.WARN))
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_data.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_WELLFORMED.format(details.status))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR11',
+                                         severity=Severity.WARNING))
 
-    def test_manifest_nomd(self):
+    def test_str12_norepmets(self):
         # test as root
-        man_no_md = STRUCT.StructureChecker('no_md', has_md=False)
-        val_errors = man_no_md.validate_manifest()
-        self.assertTrue(len(val_errors) == 1)
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR5',
-                                         severity=Severity.WARN))
-        val_errors = man_no_md.validate_manifest(is_root=False)
-        self.assertTrue(len(val_errors) == 2)
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR13',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR11',
-                                         severity=Severity.WARN))
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_repmets.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_WELLFORMED.format(details.status))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR12',
+                                         severity=Severity.WARNING))
 
-    def test_manifest_noschema(self):
+    def test_str13_norepmd(self):
         # test as root
-        man_no_schema = STRUCT.StructureChecker('no_schema', has_schema=False)
-        val_errors = man_no_schema.validate_manifest()
-        self.assertTrue(len(val_errors) == 1)
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR15',
-                                         severity=Severity.WARN))
-        val_errors = man_no_schema.validate_manifest(is_root=False)
-        self.assertTrue(len(val_errors) == 2)
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR15',
-                                         severity=Severity.WARN))
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR11',
-                                         severity=Severity.WARN))
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_repmd.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_WELLFORMED.format(details.status))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR13',
+                                         severity=Severity.WARNING))
 
-    def test_manifest_data(self):
+    def test_str15_noschema(self):
         # test as root
-        man_data = STRUCT.StructureChecker('data', has_data=True)
-        val_errors = man_data.validate_manifest()
-        self.assertTrue(len(val_errors) == 0)
-        val_errors = man_data.validate_manifest(is_root=False)
-        self.assertTrue(len(val_errors) == 0)
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_schemas.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
 
-    def test_manifest_noreps(self):
-        man_no_reps = STRUCT.StructureChecker('no_reps', has_reps=False)
-        val_errors = man_no_reps.validate_manifest()
-        self.assertTrue(len(val_errors) == 1)
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR9',
-                                         severity=Severity.WARN))
-        val_errors = man_no_reps.validate_manifest(is_root=False)
-        self.assertTrue(len(val_errors) == 1)
-        self.assertTrue(contains_rule_id(val_errors, 'CSIPSTR11',
-                                         severity=Severity.WARN))
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_WELLFORMED.format(details.status))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR15',
+                                         severity=Severity.WARNING))
+
+    def test_str16_nodocs(self):
+        # test as root
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_docs.tar.gz'))
+        _, details = STRUCT.validate(ip_path)
+        self.assertEqual(details.status, STRUCT.StructureStatus.WELLFORMED,
+                        EXP_WELLFORMED.format(details.status))
+        warn_count = 1
+        self.assertEqual(len(details.messages), warn_count,
+                        EXP_ERRORS.format(warn_count, len(details.warnings)))
+        self.assertTrue(contains_rule_id(details.warnings, 'CSIPSTR16',
+                                         severity=Severity.WARNING))
+
+    def test_get_reps(self):
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_messages.tar.gz'))
+        checker: STRUCT.StructureChecker = STRUCT.StructureChecker(ip_path)
+        reps_count = 1
+        self.assertEqual(len(checker.get_representations()), reps_count,
+                        EXP_ERRORS.format(reps_count, len(checker.get_representations())))
+
+    def test_get_no_reps(self):
+        ip_path = Path(os.path.join(self.ip_res_root, 'struct',
+                               'no_reps.tar.gz'))
+        checker: STRUCT.StructureChecker = STRUCT.StructureChecker(ip_path)
+        reps_count = 0
+        self.assertEqual(len(checker.get_representations()), reps_count,
+                        EXP_ERRORS.format(reps_count, len(checker.get_representations())))
