@@ -23,133 +23,49 @@
 # under the License.
 #
 """Module covering information package structure validation and navigation."""
-from enum import Enum, unique
-from lxml import etree as ET
 import os
+from enum import Enum, unique
+from typing import Optional
 
 from importlib_resources import files
+from lxml import etree as ET
 
-from eark_validator.ipxml import PROFILES
-from eark_validator.ipxml.schema import METS_PROF_SCHEMA
+from eark_validator.const import NO_PATH, NOT_FILE
 from eark_validator.ipxml.namespaces import Namespaces
-from eark_validator.specifications.struct_reqs import STRUCT_REQS
-from eark_validator.const import NOT_FILE, NO_PATH
+from eark_validator.ipxml.resources import profiles
+from eark_validator.ipxml.schema import METS_PROF_SCHEMA
+from eark_validator.model.specifications import Requirement, Specification
+from eark_validator.specifications.struct_reqs import REQUIREMENTS
+from eark_validator.specifications.struct_reqs import Level
 
-class Specification:
-    """Stores the vital facts and figures an IP specification."""
-    def __init__(self, title: str, url: str, version: str, date: str, requirements:dict[str, 'Requirement']=None):
-        self._title = title
-        self._url = url
-        self._version = version
-        self._date = date
-        self._requirements = requirements if requirements else {}
 
-    @property
-    def id(self) -> str:
-        """Get the id of the specification."""
-        return EarkSpecifications.from_id(self.url).name
-
-    @property
-    def title(self) -> str:
-        """Get the name of the specification."""
-        return self._title
-
-    @property
-    def url(self) -> str:
-        """Get the name of the specification."""
-        return self._url
-
-    @property
-    def version(self) -> str:
-        """Get the version."""
-        return self._version
-
-    @property
-    def date(self) -> str:
-        """Return the specification date."""
-        return self._date
-
-    @property
-    def requirements(self):
-        """Get the specification rules."""
-        for section in self.sections:
-            for requirement in self._requirements[section].values():
-                yield requirement
-
-    @property
-    def requirement_count(self) -> int:
-        """Return the number of requirments in the specification."""
-        req_count = 0
-        for sect in self.sections:
-            req_count += len(self._requirements[sect])
-        return req_count
-
-    def get_requirement_by_id(self, id: str) -> 'Requirement':
-        """Retrieve a requirement by id."""
-        for sect in self.sections:
-            req = self.get_requirement_by_sect(id, sect)
-            if req:
-                return req
-        return None
-
-    def get_requirement_by_sect(self, id: str, section: str) -> 'Requirement':
-        """Retrieve a requirement by id."""
-        sect = self._requirements[section]
-        if sect:
-            return sect.get(id)
-        return None
-
-    def section_requirements(self, section: str=None) -> list['Requirement']:
-        """Get the specification requirements, by section if offered."""
-        requirements = []
-        if section:
-            requirements = self._requirements[section]
-        else:
-            for sect in self.sections:
-                requirements += self._requirements[sect].values()
-        return requirements
-
-    @property
-    def section_count(self) -> int:
-        """Get the specification sections."""
-        return len(self._requirements)
-
-    @property
-    def sections(self) -> list[str]:
-        """Get the specification sections."""
-        return self._requirements.keys()
-
-    def __str__(self) -> str:
-        return 'name:' + self.title + ', version:' + \
-            str(self.version) + ', date:' + str(self.date)
+class Specifications:
 
     @classmethod
-    def _from_xml_file(cls, xml_file: str, add_struct: bool=False) -> 'Specification':
+    def _from_xml_file(cls, xml_file: str) -> Specification:
+        """Create a Specification from an XML file."""
         if not os.path.exists(xml_file):
             raise FileNotFoundError(NO_PATH.format(xml_file))
         if not os.path.isfile(xml_file):
             raise ValueError(NOT_FILE.format(xml_file))
-        """Create a Specification from an XML file."""
         tree = ET.parse(xml_file, parser=cls._parser())
-        return  cls._from_xml(tree, add_struct=add_struct)
+        return cls._from_xml(tree)
 
     @classmethod
     def _parser(cls) -> ET.XMLParser:
         """Create a parser for the specification."""
-        parser = ET.XMLParser(schema=METS_PROF_SCHEMA, resolve_entities=False, no_network=True)
-        return parser
+        return ET.XMLParser(schema=METS_PROF_SCHEMA, resolve_entities=False, no_network=True)
 
     @classmethod
-    def _from_xml(cls, tree: ET.ElementTree, add_struct: bool=False) -> 'Specification':
-        spec = cls.from_element(tree.getroot(), add_struct=add_struct)
-        return spec
+    def _from_xml(cls, tree: ET.ElementTree) -> Specification:
+        return cls.from_element(tree.getroot())
 
     @classmethod
-    def from_element(cls, spec_ele: ET.Element, add_struct: bool=False) -> 'Specification':
+    def from_element(cls, spec_ele: ET.Element) -> Specification:
         """Create a Specification from an XML element."""
         version = spec_ele.get('ID')
         title = date = ''
-        requirements = {}
+        requirements: dict[str, Requirement] = {}
         profile = ''
         # Loop through the child eles
         for child in spec_ele:
@@ -163,170 +79,120 @@ class Specification:
                 requirements = cls._processs_requirements(child)
             elif child.tag in [Namespaces.PROFILE.qualify('URI'), 'URI']:
                 profile = child.text
-        if add_struct:
-            # Add the structural requirements
-            struct_reqs = Specification.StructuralRequirement._get_struct_reqs()
-            requirements['structure'] = struct_reqs
+        # Add the structural requirements
+        struct_reqs = StructuralRequirements.get_requirements()
         # Return the Specification
-        return cls(title, profile, version, date, requirements=requirements)
+        return Specification.model_validate({
+            'title': title,
+            'url': profile,
+            'version': version,
+            'date': date,
+            'requirements': requirements,
+            'structural_requirements': struct_reqs
+            })
 
     @classmethod
     def _processs_requirements(cls, req_root: ET.Element) -> dict[str, 'Requirement']:
         requirements = {}
         for sect_ele in req_root:
             section = sect_ele.tag.replace(Namespaces.PROFILE.qualifier, '')
-            reqs = {}
+            reqs = []
             for req_ele in sect_ele:
-                requirement = cls.Requirement.from_element(req_ele)
+                requirement = Requirements.from_element(req_ele)
                 if not requirement.id.startswith('REF_'):
-                    reqs.update({requirement.id: requirement})
+                    reqs.append(requirement)
             requirements[section] = reqs
         return requirements
 
-    class Requirement():
-        """Encapsulates a requirement."""
-        def __init__(self, req_id: str, name: str, level: str='MUST', xpath: str=None, cardinality: str=None):
-            self._id = req_id
-            self._name = name
-            self._level = level
-            self._xpath = xpath
-            self._cardinality = cardinality
+class Requirements():
+    @staticmethod
+    def from_element(req_ele: ET.Element) -> Requirement:
+        """Return a Requirement instance from an XML element."""
+        req_id = req_ele.get('ID')
+        level: Level = Level.from_string(req_ele.get('REQLEVEL'))
+        name = ''
+        for child in req_ele:
+            if child.tag == Namespaces.PROFILE.qualify('description'):
+                for req_child in child:
+                    if req_child.tag == Namespaces.PROFILE.qualify('head'):
+                        name = req_child.text
+        return Requirement.model_validate({
+            'id': req_id,
+            'name': name,
+            'level': level
+            })
 
-        @property
-        def id(self) -> str: # pylint: disable-msg=C0103
-            """Return the id."""
-            return self._id
+class StructuralRequirements():
+    @staticmethod
+    def from_rule_no(rule_no: int) -> Requirement:
+        """Create an StructuralRequirement from a numerical rule id and a sub_message."""
+        item = REQUIREMENTS.get(rule_no)
+        if not item:
+            raise ValueError(f'No rule with number {rule_no}')
+        return StructuralRequirements.from_dictionary(item)
 
-        @property
-        def name(self) -> str:
-            """Return the name."""
-            return self._name
+    @staticmethod
+    def from_dictionary(item: dict[str, str]) -> Requirement:
+        """Create an StructuralRequirement from dictionary item and a sub_message."""
+        return Requirement.model_validate({
+                'id': item.get('id'),
+                'level': item.get('level'),
+                'message': item.get('message')
+            })
 
-        @property
-        def level(self) -> str:
-            """Return the level."""
-            return self._level
-
-        @property
-        def xpath(self) -> str:
-            """Return the xpath."""
-            return self._xpath
-
-        @property
-        def cardinality(self) -> str:
-            """Return the cardinality."""
-            return self._cardinality
-
-        def __str__(self) -> str:
-            return 'id:' + self.id + ', name:' + self.name
-
-        @classmethod
-        def from_element(cls, req_ele: ET.Element) -> 'Specification.Requirement':
-            """Return a Requirement instance from an XML element."""
-            req_id = req_ele.get('ID')
-            level = req_ele.get('LEVEL')
-            name = ''
-            for child in req_ele:
-                if child.tag == Namespaces.METS.qualify('description'):
-                    for req_child in child:
-                        if req_child.tag == Namespaces.METS.qualify('head'):
-                            name = req_child.text
-            return cls(req_id, name, level)
-
-    class StructuralRequirement():
-        """Encapsulates a structural requirement."""
-        def __init__(self, req_id: str, level: str='MUST', message: str=None):
-            self._id = req_id
-            self._level = level
-            self._message = message
-
-        @property
-        def id(self) -> str: # pylint: disable-msg=C0103
-            """Return the id."""
-            return self._id
-
-        @property
-        def level(self) -> str:
-            """Return the level."""
-            return self._level
-
-        @property
-        def message(self) -> str:
-            """Return the message."""
-            return self._message
-
-        def __str__(self) -> str:
-            return 'id:' + self.id + ', level:' + str(self.level)
-
-        @classmethod
-        def from_rule_no(cls, rule_no: int) -> 'Specification.StructuralRequirement':
-            """Create an StructuralRequirement from a numerical rule id and a sub_message."""
-            item = STRUCT_REQS.get(rule_no)
-            return cls.from_dict_item(item)
-
-        @classmethod
-        def from_dict_item(cls, item: ET.Element) -> 'Specification.StructuralRequirement':
-            """Create an StructuralRequirement from dictionary item and a sub_message."""
-            return cls.from_values(item.get('id'), item.get('level'),
-                                   item.get('message'))
-
-        @classmethod
-        def from_values(cls, req_id: str, level: str='MUST', message:str=None) -> 'Specification.StructuralRequirement':
-            """Create an StructuralRequirement from values supplied."""
-            return cls(req_id, level, message)
-
-        @staticmethod
-        def _get_struct_reqs() -> list['Specification.StructuralRequirement']:
-            reqs = []
-            for req_num in STRUCT_REQS:
-                req = STRUCT_REQS.get(req_num)
-                reqs.append(Specification.StructuralRequirement(req.get('id'),
-                                                                level=req.get('level'),
-                                                                message=req.get('message')))
-            return reqs
-
+    @staticmethod
+    def get_requirements() -> list[Requirement]:
+        reqs = []
+        for req in REQUIREMENTS.values():
+            reqs.append(Requirement.model_validate(req))
+        return reqs
 
 @unique
-class EarkSpecifications(Enum):
-    """Enumeration of E-ARK specifications."""
+class SpecificationVersion(str, Enum):
+    V2_0_4 = 'V2.0.4'
+    V2_1_0 = 'V2.1.0'
+
+    def __str__(self):
+        return self.value
+
+@unique
+class SpecificationType(str, Enum):
     CSIP = 'E-ARK-CSIP'
     SIP = 'E-ARK-SIP'
     DIP = 'E-ARK-DIP'
 
-    def __init__(self, value: str):
-        self._path = str(files(PROFILES).joinpath(value + '.xml'))
-        self._specfication = Specification._from_xml_file(self._path)
-        self._title = value
+    @classmethod
+    def from_string(cls, type: str) -> Optional['SpecificationType']:
+        """Get the enum from the value."""
+        for spec in cls:
+            if type in [spec.name, spec.value]:
+                return spec
+        raise ValueError('{type} does not exists')
+
+class EarkSpecification:
+    def __init__(self, type: SpecificationType, version: SpecificationVersion):
+        self._type: SpecificationType = type
+        self._version: SpecificationVersion = version
+
+        self._path = str(files(profiles).joinpath(version).joinpath(type + '.xml'))
+        self._specfication = Specifications._from_xml_file(self.path)
 
     @property
-    def id(self) -> str:
-        """Get the specification id."""
-        return self.name
+    def version(self) -> SpecificationVersion:
+        """Get the specification version."""
+        return self._version
+
+    @property
+    def type(self) -> SpecificationType:
+        """Get the specification type."""
+        return self._type
 
     @property
     def path(self) -> str:
         """Get the path to the specification file."""
-        self._path
-
-    @property
-    def title(self) -> str:
-        """Get the specification title."""
-        self._title
+        return self._path
 
     @property
     def specification(self) -> Specification:
         """Get the specification."""
         return self._specfication
-
-    @property
-    def profile(self) -> str:
-        """Get the specification profile url."""
-        return 'https://eark{}.dilcis.eu/profile/{}.xml'.format(self.name.lower(), self.value)
-
-    @classmethod
-    def from_id(cls, id: str) -> 'EarkSpecifications':
-        """Get the enum from the value."""
-        for spec in cls:
-            if spec.id == id or spec.value == id or spec.profile == id:
-                return spec
-        return None
