@@ -22,19 +22,26 @@
 # specific language governing permissions and limitations
 # under the License.
 #
+from typing import List
 import unittest
 
 from enum import Enum
 
 from importlib_resources import files
+from pydantic import ValidationError
 
 from eark_validator import rules as SC
-from eark_validator.specifications.specification import EarkSpecifications
+from eark_validator.model.validation_report import Severity, Result, ValidationReport
+from eark_validator.specifications.specification import SpecificationType, SpecificationVersion
 import tests.resources.schematron as SCHEMATRON
 import tests.resources.xml as XML
+import tests.resources.json as JSON
 
 TEST_RES = 'tests.resources'
 TEST_RES_XML = TEST_RES + '.xml'
+TEST_RES_JSON = TEST_RES + '.json'
+COMMONS_IP_JSON = str(files(JSON).joinpath('commons-ip-report.json'))
+COMMONS_IP_INVALID_JSON = str(files(JSON).joinpath('commons-ip-invalid.json'))
 PERSON_PATH = str(files(SCHEMATRON).joinpath('person.xml'))
 NOT_FOUND_PATH = str(files(SCHEMATRON).joinpath('not-found.xml'))
 EMPTY_FILE_PATH = str(files(TEST_RES).joinpath('empty.file'))
@@ -69,8 +76,8 @@ class ValidationRulesTest(unittest.TestCase):
 
     def test_mets_root_other_type(self):
         result, failures, _, _ = _test_validation(METS_ROOT_RULES, 'METS-other-type.xml')
-        self.assertEqual(failures, 0)
-        self.assertTrue(result)
+        self.assertEqual(failures, 1)
+        self.assertFalse(result)
 
     def test_mets_root_no_profile(self):
         result, failures, _, _ = _test_validation(METS_ROOT_RULES, 'METS-no-profile.xml')
@@ -108,10 +115,10 @@ class ValidationRulesTest(unittest.TestCase):
         self.assertFalse(result)
 
     def test_mets_root_dmd(self):
-        result, _, warnings = _full_validation(METS_ROOT_RULES, METS_VALID)
+        results = _full_validation(METS_ROOT_RULES, METS_VALID)
         found_csip17 = False
-        for warning in warnings:
-            if warning.rule_id == 'CSIP17':
+        for result in results:
+            if result.rule_id == 'CSIP17':
                 found_csip17 = True
         self.assertTrue(found_csip17)
         self.assertTrue(result)
@@ -137,141 +144,137 @@ class ValidationRulesTest(unittest.TestCase):
     def test_mets_structmap(self):
         result, failures, warnings, _ = _test_validation(METS_STRUCT_RULES, METS_VALID)
         self.assertEqual(failures, 0)
-        self.assertEqual(warnings, 1)
+        self.assertEqual(warnings, 3)
         self.assertTrue(result)
 
 class ValidationProfileTest(unittest.TestCase):
     def test_load_by_str(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         self.assertEqual(profile.specification.url, CSIP_PROF)
-        profile = SC.ValidationProfile.from_specification('SIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('SIP'), SpecificationVersion.V2_0_4)
         self.assertEqual(profile.specification.url, SIP_PROF)
-        profile = SC.ValidationProfile.from_specification('DIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('DIP'), SpecificationVersion.V2_0_4)
         self.assertEqual(profile.specification.url, DIP_PROF)
 
     def test_load_by_eark_spec(self):
-        profile = SC.ValidationProfile.from_specification(EarkSpecifications.CSIP)
+        profile = SC.ValidationProfile(SpecificationType.CSIP, SpecificationVersion.V2_0_4)
         self.assertEqual(profile.specification.url, CSIP_PROF)
-        profile = SC.ValidationProfile.from_specification(EarkSpecifications.SIP)
+        profile = SC.ValidationProfile(SpecificationType.SIP, SpecificationVersion.V2_0_4)
         self.assertEqual(profile.specification.url, SIP_PROF)
-        profile = SC.ValidationProfile.from_specification(EarkSpecifications.DIP)
-        self.assertEqual(profile.specification.url, DIP_PROF)
-
-    def test_load_by_spec(self):
-        profile = SC.ValidationProfile.from_specification(EarkSpecifications.CSIP.specification)
-        self.assertEqual(profile.specification.url, CSIP_PROF)
-        profile = SC.ValidationProfile.from_specification(EarkSpecifications.SIP.specification)
-        self.assertEqual(profile.specification.url, SIP_PROF)
-        profile = SC.ValidationProfile.from_specification(EarkSpecifications.DIP.specification)
+        profile = SC.ValidationProfile(SpecificationType.DIP, SpecificationVersion.V2_0_4)
         self.assertEqual(profile.specification.url, DIP_PROF)
 
     def test_bad_value(self):
         with self.assertRaises(ValueError):
-            SC.ValidationProfile.from_specification('BAD')
+            SC.ValidationProfile(SpecificationType.from_string('BAD'), SpecificationVersion.V2_0_4)
+    def test_unimplemented_specifications(self):
+        with self.assertRaises(ValueError):
+            SC.ValidationProfile(SpecificationType.from_string('AIP'), SpecificationVersion.V2_0_4)
+        with self.assertRaises(ValueError):
+            SC.ValidationProfile(SpecificationType.from_string('AIU'), SpecificationVersion.V2_0_4)
+        with self.assertRaises(ValueError):
+            SC.ValidationProfile(SpecificationType.from_string('AIC'), SpecificationVersion.V2_0_4)
 
     def test_valid(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES_XML).joinpath(METS_VALID)))
         self.assertTrue(profile.is_valid)
 
     def test_invalid(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES_XML).joinpath('METS-no-hdr.xml')))
         self.assertFalse(profile.is_valid)
 
     def test_validate_file_not_found(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         with self.assertRaises(FileNotFoundError):
             profile.validate(str(files(SCHEMATRON).joinpath('not-found.xml')))
 
     def test_validate_dir_value_err(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         with self.assertRaises(ValueError):
             profile.validate(str(files(SCHEMATRON)))
 
     def test_validate_empty_file(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES).joinpath('empty.file')))
         self.assertFalse(profile.is_valid)
 
     def test_validate_not_mets(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES_XML).joinpath('person.xml')))
         self.assertFalse(profile.is_valid)
 
     def test_validate_json(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES).joinpath('aip.json')))
         self.assertFalse(profile.is_valid)
 
     def test_get_results(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES_XML).joinpath(METS_VALID)))
         self.assertTrue(profile.is_valid)
         self.assertEqual(len(profile.get_results()), 8)
 
     def test_get_result(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES_XML).joinpath(METS_VALID)))
         result = profile.get_result('metsHdr')
         self.assertTrue(profile.is_valid)
-        self.assertEqual(len(result.warnings), 1)
+        self.assertEqual(len(list(filter(lambda a: a.severity == Severity.WARNING, result))), 1)
 
     def test_get_bad_key(self):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.from_string('CSIP'), SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES_XML).joinpath(METS_VALID)))
         result = profile.get_result('badkey')
         self.assertIsNone(result)
 
-class SeverityTest(Enum):
-    NOT_SEV: 'NOT_SEV'
+    def test_deserialise_commons_ip_report(self):
+        file_name = COMMONS_IP_JSON
+        with open(file_name, 'r', encoding='utf-8') as _f:
+            contents = _f.read()
+        result: ValidationReport = ValidationReport.model_validate_json(contents)
+        self.assertIsNotNone(result)
+
+    def test_deserialise_commons_ip_invalid(self):
+        file_name = COMMONS_IP_INVALID_JSON
+        with open(file_name, 'r', encoding='utf-8') as _f:
+            contents = _f.read()
+        result: ValidationReport = ValidationReport.model_validate_json(contents)
+        self.assertIsNotNone(result)
+
+class SeverityTest(str, Enum):
+    NOT_SEV = 'NOT_SEV'
 
 class ResultTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        profile = SC.ValidationProfile.from_specification('CSIP')
+        profile = SC.ValidationProfile(SpecificationType.CSIP, SpecificationVersion.V2_0_4)
         profile.validate(str(files(TEST_RES_XML).joinpath(METS_VALID)))
-        cls._result = profile.get_result('metsHdr').warnings[0]
+        cls._result = profile.get_result('metsHdr')[0]
 
     def test_get_message(self):
         self.assertIsNotNone(self._result.message)
 
-    def test_sev_instances(self):
-        for sev in SC.Severity:
-            self._result.severity = sev
-
-    def test_sev_names(self):
-        for sev in SC.Severity:
-            self._result.severity = sev.name
-
-    def test_sev_values(self):
-        for sev in SC.Severity:
-            self._result.severity = sev.value
-
-    def test_bad_sev_string(self):
-        with self.assertRaises(ValueError):
-            self._result.severity = 'BAD'
-
     def test_bad_sev_att(self):
-        with self.assertRaises(AttributeError):
-            self._result.severity = SeverityTest.NOT_SEV
+        with self.assertRaises(ValidationError):
+            Result.model_validate({ 'severity': SeverityTest.NOT_SEV })
 
 def _test_validation(name, to_validate):
-    rules = SC.SchematronRuleset(SC.get_schematron_path('CSIP', name))
+    rules = SC.SchematronRuleset(SC.get_schematron_path(SpecificationVersion.V2_0_4, 'CSIP', name))
     rules.validate(str(files(XML).joinpath(to_validate)))
-    for failure in SC.TestReport.from_validation_report(rules._schematron.validation_report).errors:
-        print(failure)
-    for warning in SC.TestReport.from_validation_report(rules._schematron.validation_report).warnings:
-        print(warning)
-    report = SC.TestReport.from_validation_report(rules._schematron.validation_report)
-    return report.is_valid, len(report.errors), len(report.warnings), len(report.infos)
+    results: List[Result] = SC.TestResults.from_validation_report(rules._schematron.validation_report)
+    errors = warnings = infos = 0
+    for result in results:
+        if result.severity == SC.Severity.ERROR:
+            errors += 1
+        elif result.severity == SC.Severity.WARNING:
+            warnings += 1
+        elif result.severity == SC.Severity.INFORMATION:
+            infos += 1
+    return errors < 1, errors, warnings, infos
 
 def _full_validation(name, to_validate):
-    rules = SC.SchematronRuleset(SC.get_schematron_path('CSIP', name))
+    rules = SC.SchematronRuleset(SC.get_schematron_path(SpecificationVersion.V2_0_4, 'CSIP', name))
     rules.validate(str(files(XML).joinpath(to_validate)))
-    for failure in SC.TestReport.from_validation_report(rules._schematron.validation_report).errors:
-        print(failure)
-    for warning in SC.TestReport.from_validation_report(rules._schematron.validation_report).warnings:
-        print(warning)
-    report = SC.TestReport.from_validation_report(rules._schematron.validation_report)
-    return report.is_valid, report.errors, report.warnings
+    return SC.TestResults.from_validation_report(rules._schematron.validation_report)
