@@ -23,11 +23,11 @@
 # under the License.
 #
 """Module that holds common utilities for unit testing."""
-from eark_validator.rules import Severity
+from eark_validator.model import Severity
 def contains_rule_id(error_list, rule_id, severity=Severity.ERROR):
     """Check that a particular error with specified severity is present in a list
     of errors."""
     for val_error in error_list:
-        if val_error.id == rule_id and val_error.severity == severity:
+        if val_error.rule_id == rule_id and (val_error.severity is severity or val_error.severity == severity.value)  :
             return True
     return False
